feat(308): migrate app state to the new architecture
