+++ features/core/state/index.ts
@@ -1,0 +1,10 @@
+export { useAppState } from '@/features/core/state/hooks/useAppState';
+export { useModalState } from '@/features/core/state/hooks/useModalState';
+export { createSelectors } from '@/features/core/state/libraries/createSelectors';
+export { create, resetAllStores } from '@/features/core/state/libraries/createStore';
+
+export type {
+  ActionModalContentType,
+  InfoModalContentType,
+  ResetPasswordModalContentType,
+} from '@/features/core/state/hooks/modalStore';
