+++ features/core/state/libraries/createSelectors.ts
@@ -1,0 +1,13 @@
+import type { StoreApi, UseBoundStore } from 'zustand';
+
+type WithSelectors<S> = S extends { getState: () => infer T } ? S & { use: { [K in keyof T]: () => T[K] } } : never;
+
+export const createSelectors = <S extends UseBoundStore<StoreApi<Record<string, unknown>>>>(_store: S) => {
+  const store = _store as WithSelectors<typeof _store>;
+  store.use = {};
+  for (const k of Object.keys(store.getState())) {
+    (store.use as Record<string, () => unknown>)[k] = () => store(s => s[k]);
+  }
+
+  return store;
+};
