--- ui/state/app/index.ts
@@ -1,1 +1,0 @@
-export { useAppState } from './useAppState';
