--- ui/state/shared/createSelectors.ts
@@ -1,13 +1,0 @@
-import type { StoreApi, UseBoundStore } from 'zustand';
-
-import type { WithSelectors } from './types';
-
-export const createSelectors = <S extends UseBoundStore<StoreApi<object>>>(_store: S) => {
-  const store = _store as WithSelectors<typeof _store>;
-  store.use = {};
-  for (const k of Object.keys(store.getState())) {
-    (store.use as any)[k] = () => store(s => s[k as keyof typeof s]);
-  }
-
-  return store;
-};
