--- ui/state/shared/types.ts
@@ -1,3 +1,0 @@
-export type WithSelectors<S> = S extends { getState: () => infer T }
-  ? S & { use: { [K in keyof T]: () => T[K] } }
-  : never;
